+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.14.4
+## 3.4.15.0
 
 ### New Features
 - Allow skipping browser download mods for cases where mods no longer exist